git: move actions
