+++ completers/git_completer/cmd/tag.go
@@ -4,7 +4,6 @@ import (
 	"github.com/rsteube/carapace"
 	"github.com/rsteube/carapace-bin/pkg/actions/os"
 	"github.com/rsteube/carapace-bin/pkg/actions/tools/git"
-	"github.com/rsteube/carapace/pkg/style"
 	"github.com/spf13/cobra"
 )
 
@@ -47,7 +46,7 @@ func init() {
 
 	carapace.Gen(tagCmd).FlagCompletion(carapace.ActionMap{
 		"cleanup":     git.ActionCleanupMode(),
-		"color":       carapace.ActionValues("auto", "never", "always").StyleF(style.ForKeyword),
+		"color":       git.ActionColorModes(),
 		"contains":    git.ActionRefs(git.RefOption{Commits: 100, HeadCommits: 100}),
 		"file":        carapace.ActionFiles(),
 		"local-user":  os.ActionGpgKeyIds(),
