+++ pkg/actions/tools/git/color.go
@@ -1,0 +1,14 @@
+package git
+
+import (
+	"github.com/rsteube/carapace"
+	"github.com/rsteube/carapace/pkg/style"
+)
+
+// ActionColorModes completes color modes
+//
+//	always
+//	auto
+func ActionColorModes() carapace.Action {
+	return carapace.ActionValues("always", "auto", "never").StyleF(style.ForKeyword)
+}
